syamsul/update-grule-service

+++ internal/model/discount.go
@@ -3,5 +3,5 @@ package model
 type Discount struct {
 	isContinentSoldOut bool
 	isContinentDiff    bool
-	DiscountAmount     int
+	DiscountAmount     float32
 }
